mimir/requirements.txt
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
defusedxml==0.5.0
Django==2.0.4
django-bower==5.2.0
idna==2.6
jieba==0.39
networkx==2.1
nltk==3.2.5
numpy==1.14.2
oauthlib==2.0.7
pandas==0.22.0
PyJWT==1.6.1
pymongo==3.6.1
python-dateutil==2.7.2
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.22
